rename source files for compute pentropy/atom to match the command name
