+++ src/USER-MISC/compute_pentropy_atom.h
